refactor: move types to separate file
